refactor(paper): extract submission and reconciliation services
