+++ src/marketlab/paper/application/__init__.py
@@ -1,7 +1,11 @@
 from .approval import ApprovalService
 from .decision import DecisionService
+from .reconciliation import ReconciliationService
+from .submission import SubmissionService
 
 __all__ = [
     "ApprovalService",
     "DecisionService",
+    "ReconciliationService",
+    "SubmissionService",
 ]
